restructured components
