+++ apps/storybook/stories/components/Layout/Center/Center.stories.tsx
@@ -1,4 +1,4 @@
-import { Center as KvibCenter } from "@kvib/react/src/components/layout";
+import { Center as KvibCenter } from "@kvib/react/src/layout";
 import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof KvibCenter> = {
