+++ apps/storybook/stories/components/Layout/Container/Container.stories.tsx
@@ -1,4 +1,4 @@
-import { Container as KvibContainer } from "@kvib/react/src/components/layout";
+import { Container as KvibContainer } from "@kvib/react/src/layout";
 import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof KvibContainer> = {
