+++ apps/storybook/stories/components/Layout/Flex/Flex.stories.tsx
@@ -1,4 +1,4 @@
-import { Center, Box, Flex as KvibFlex, Spacer } from "@kvib/react/src/components/layout";
+import { Center, Box, Flex as KvibFlex, Spacer } from "@kvib/react/src/layout";
 import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof KvibFlex> = {
