+++ apps/storybook/stories/components/Layout/SimpleGrid/SimpleGrid.stories.tsx
@@ -1,4 +1,4 @@
-import { SimpleGrid as KvibSimpleGrid, Box } from "@kvib/react/src/components/layout";
+import { SimpleGrid as KvibSimpleGrid, Box } from "@kvib/react/src/layout";
 import { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof KvibSimpleGrid> = {
