--- packages/react/src/components/index.tsx
@@ -1,3 +1,0 @@
-export * from "./button";
-export * from "./layout";
-export * from "./provider";
