+++ packages/react/src/index.ts
@@ -1,2 +1,4 @@
-export * from "./components";
+export * from "./provider";
 export * from "./theme";
+export * from "./button";
+export * from "./layout";
